requirements.txt
crewai>=0.11.0
python-dotenv>=0.19.0
requests>=2.31.0
feedparser>=6.0.10 